requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
aiomysql==0.2.0
pymysql==1.1.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
firebase-admin==6.4.0
pydantic-settings==2.1.0
httpx==0.26.0
boto3==1.34.34
aioboto3==12.3.0
GeoAlchemy2==0.14.3
fastapi-mail==1.4.1
aiosmtplib==2.0.2